Use the subcycling-enabled retry CTU advance for simplified SDC (#695)
The retry mechanism can be used for the simplified SDC advance as well as the standard CTU advance.

I've made some changes to improve this and make it quite simple. Now we simply do the SDC iterations inside the outer while loop that does retries.

I've also made some cosmetic changes to the return values of the CTU advance and react_state to return bools, indicating whether or not they were successful (unsuccessful meaning we either abort or retry).