Add an option for implicit external source terms (#709)
The external source terms are currently only implemented as an explicit predictor-corrector scheme. This is incompatible with how wdmerger implements its relaxation source term, so wdmerger is currently not working as intended. This PR introduces a new option, castro.ext_src_implicit (defaulting to 0) which, if turned on, changes the handling of the external source terms to allow an implicit solve. This is done by subtracting the full old-time source and adding the full new-time source in the corrector, rather than -0.5 and +0.5 of each, respectively. It would still be up to the individual problem to make sure they are consistent with this scheme if they turn the option on. Even if they don't, the source term ends up being a simple time-advanced source term at n+1 which is not a serious problem.