Add option to apply sources consecutively (#710)
This new option allows you to add sources to the state as they are calculated, rather than grouping them together at the end. Its purpose is to allow sources that come at the end, in particular the sponge and the external source term, to act on the state after the actual physics source terms have already been applied.