Use existing integrate_grav for make_radial_phi (#717)
make_radial_phi had a simple approach for integrating the mass distribution to get the gravitational acceleration, which was then integrated to get the gravitational potential. This PR uses the existing ca_integrate_grav to do the first step. Since the algorithm is different (more sophisticated) answers will change. This step is in preparation for offloading the radial gravity routines.