Store RadSolve in Castro instead of creating on demand
The prior approach of creating a RadSolve object on demand allowed for
maximum flexibility but is not optimal for performance, since creation
and destruction of all of the array data is very expensive, especially
on GPUs (relative to the compute phase). Now we store the RadSolve
object in Castro, similar to other Castro-owned data, and it is filled
on Castro construction (which occurs at initialization or after a regrid).
As a result, for a single-level solve (or a multi-level solve with no
regrids), we only pay the cost of array initialization once, allowing us
to amortize the cost over many timesteps and effectively make it free.