Remove self-gravity define, use GRAVITY always (#719)
This PR closes #157 by removing the SELF_GRAVITY ifdef -- we revert back to the previous behavior where GRAVITY is the only relevant define.

Also, I'm including a centralization on the approach for grav_source_type = 4 based on g, not phi. The mechanism based on phi is slightly better at conservation in the presence of coarse-fine boundaries, but the difference between the two methods is quite small in absolute terms -- both are very good -- and it is better to only have one method so that we don't have to worry when switching between different gravity implementations (not all of them define phi).