Set compute_new_dt_on_regrid by default (#720)
The AMR parameter amr.compute_new_dt_on_regrid is now on by default; closes #356. This avoids crashes that result from the CFL number being too large after regridding, because we update the timestep after seeing that larger velocity.

The user can still opt to set this off if they want to, since the location where we're setting it in Castro occurs before the Amr ParmParse of this parameter.