Replace do_real_eos = 0 with an actual EOS (#722)
This PR adds an EOS, rad_power_law, which replaces the functionality previously provided by radiation.do_real_eos = 0. The EOS defines c_v as before, and can only be called in (rho, T) and (rho, e) modes. It does not define hydrodynamic variables like pressure, so it cannot be used in conjunction with castro.do_hydro = 1. This sets us up for merging the radiation and non-radiation EOS interfaces.