Remove integrate_Planck option, always integrate by default (#740)
There existed an option to turn off the normal behavior of integrating
the Planck function to get the blackbody spectrum at each group center.
This was only used in the RadSphere test, and the alternative was to
evalute the Planck function as piecewise constant within each energy
group. This introduces code complexity without any obvious gain; we still
get reasonable results using the more general expression, and reduce
the number of code paths to maintain.