Variable cleanup cxx (#754)
This switches the initialization of the C++ state indices (e.g. Density, QRHO, ...) to constexprs. This let's the work with the lambda-capture GPU offloading. Note: the Fortran indices are still handled the same way as before. This is an alternative to #747

As part of this change, the general_null network writes out a C++ header.