Replace Castro microphysics with starkiller-astro Microphysics (#760)
This implements the Microphysics part of #651. We replace the local Castro Microphysics directory with a dependence on the main starkiller-astro Microphysics directory. This means that obtaining Microphysics is now a requirement for using Castro. If the user sets MICROPHYSICS_HOME as before, then their workflow does not change. A new path is that Microphysics is a git submodule of Castro, so if the user does git submodule update --init --recursive then they will obtain the version of Microphysics that we specify in the repo, and if they do not have MICROPHYSICS_HOME set then this submodule will be used. The repo is now updated with git pull --recurse-submodules, so that updates to Microphysics come along when you pull.

A minor associated change is that we've standardized on NETWORK_INPUTS in all of the makefiles, instead of using GENERAL_NET_INPUTS with the hardcoded path.