Handle the plotfile case where a derived variable has multiple components (#758)
In this case, the plotfile logic would only plot the first component of the derived
variable. Now, if a derived variable has multiple components, we plot all components.
The name of each is appended with an integer variable indicating which one it is.

This issue also exists in AMReX, where this plotfile code presumably originally came from.