move the EOS reset on states outside of the trans loop (#767)
This puts the reset_edge_state_thermo calls outside of the trans loops.  Note that the reason we chose to do this is because it removes the EOS from the transverse routines, which allows us to move the transverse routines to C++ independently of the implementation of the EOS in Microphysics in C++ (which is happening over a longer timescale). However, in the long run when we have got that all done, we will want to look at moving this back to be a zone-by-zone implementation.