remove the reconstruction of gamma_e on interfaces (#780)
This drops the alternative to (rho e) on interfaces -- reconstructing gamma_e. They both are about the same, so there's no need to support both

Note: as a consequence of this, we will be removing the ppm_temp_fix = 1 option from CTU (since apparently it was never coded into the default rhoe reconstruction for some reason). See issue #781