Remove BC fill routines other than for State_Type (#815)
This PR removes all boundary fill interfaces other than hypfill and denfill (gravity, rotation, reactions, radiation). The existence of these routines creates substantial maintenance overhead relative to the value they add, since none of the problems actually use them to update the boundaries. As an example of the problem this complexity adds, not all of the routines were addressing inflow boundaries consistently (in particular, some override an inflow boundary to do first-order extrapolation so that there's meaningful data, and some don't, leading to garbage data for e.g. the reactions). The only one of these that a user would be likely to touch was the radiation fill, but we have no problem setups demonstrating this (they all use radiation.lo_bc and radiation.hi_bc instead), and it's also hard to see how a user could have effectively used the radiation fill for a multigroup problem, since you don't come in with information about which group you're in.