Switch over to the new Fab interface for BC fills (#819)
This is in preparation to use the new fab_filcc in AMReX.

generic_single_fill and generic_multi_fill are merged into one function, generic_fill, since the Fab interface has no notion of separate single and group interfaces. Similarly, the calls to denfill and hypfill are now both done in one function (though the functions are not merged, to preserve user-facing code).

The source fill routines are also merged into the generic fill routines.

Finally, instead of locally overriding the BCs in the boundary fill routine, we now globally override any inflow boundary conditions (for the non-State_Type data) to become FOEXTRAP. So this is the same result but with code that is a little simpler and will be easier to link up to AMReX later.