Remove src q (#829)
This removes the src_q multifab and instead converts the conserved state sources to primitive state sources FAB by FAB just before they are needed (in the creation of the interface states). This saves on memory from the MF as well as allows us to eliminate two entries in qaux (QDPDE and QDPDR). The trade-off is we now do an EOS call.

Answers may change because of the additional EOS call.