Always retry if the density is small or negative (#796)
The current code only initiates a retry if the density was updated by enforce_min_density from a sufficiently negative number to a sufficiently positive one. This made sense at the time of its implementation but a simpler method is now available: we check whether the minimum value of the density is less than small_dens after the advance, and if so we reject the timestep and retry the step.

The parameter castro.retry_neg_dens_factor is removed.