Provide a reason when an advance fails (#845)
The current code says "advance was unsuccessful" when it is rejected for a retry, but does not tell you why. This code adds a reason for each of the possible failure modes so that you have some idea of what went wrong.

To make the logic simpler to follow, check_timestep is moved into the main advance function.