Account for multiple edges in the flux limiters (#846)
It is possible even for limited fluxes to cause a negative density when multiple fluxes are summed, since we are only individually limiting fluxes. This PR accounts for that by increasing the effective floor for each flux by a factor of 6 (in 3D; lower in 1D/2D) to ensure that even if all six fluxes are evacuating at a large rate, the final density will still be > small_dens.