Disallow restarting from old checkpoints (#860)
We have maintained a checkpoint versioning system for a few years, and strove to maintain backward compatibility so that we could restart from an old checkpoint. This adds significant programmer overhead for dubious gain, since we are not aware of examples of people actually trying to use a new version of the code to restart from an old checkpoint. I propose removing this backward compatibility so that we are more free to make changes to StateData in the future without having to worry about corner cases in how this affects restarting.