Move burn weightings into Reactions_Type
We currently have a scheme for storing reactions weightings, which
are a measure of the number of RHS evaluations during the burn and
therefore a proxy for the difficulty of the burn. These weights were
added as separate StateData depending on the runtime option
use_custom_knapsack_weights.

This change instead places the weights directly in the Reactions_Type
StateData as a new component. Now that we always store the weights,
load balancing and other future changes that depend on the weights
are easier to manage.

The only thing we have to be careful of now is that code that previously
zeroed out all of the reactions components cannot be applied until after
the weights have been used for their intended purpose. (For load balancing
purposes, we want the minimum weight to be one, not zero.) For safety, then,
we now don't zero out the reactions components in the strang_react functions,
but instead handle the setting of those components directly in the react_state
loop.