Switch reduce_add to atomics for safety (#876)
reduce_add used a combined block level reduce plus atomics. This is generally unsafe inside an amrex::ParallelFor() or a #pragma gpu, but it works often enough that I didn't consider it a serious issue. But now I'm hitting some cases for small boxes where this is generating wrong answers so we need to fix it (see also #858). We should be switching this remaining code to C++ using ReduceOps soon (or pure atomics, where it makes sense) so this will become less of an issue. But until then we should fix the Fortran code.