Replace check_timestep with call to estDt (#880)
As noted in #879, check_timestep had a few issues, including that it didn't cover diffusion or simplified SDC. We replace that custom code with a call to estDt, which evaluates the new data. Then we can automatically get the new-time timestep limiter, and compare to see if it's within change_max of the old timestep. If not, we consider the step to have failed.