Remove an unnecessary estTimeStep call (#881)
When use_post_step_regrid was added, logic was added to force subcycling if we were just coming off of a post-timestep regrid. But now we always use the subcycling driver and we check the timestep for validity at the end of the step, so this is no longer necessary. Also, it seems to cause segfaults in certain cases, though I don't know why.