an attempt at making the left and right states clearer (#947)
There are a few issues in how the left and right states are named. In mhd_plm, Ip is used for the left state and Im for the right. These are then stored in qp and qm in the driver. But this is opposite the convention that we use in the rest of the code. To make things clearer, qp is now qleft and qm is now qright.

I propagated this naming convention into ct_upwind. There however, there was more confusion as the names uL and uR were opposite their inferred meanings, so this was also fixed.