Port MHD flattening to C++ and address box sizes (#984)
Now the flattening is done over the same size box as we do reconstruction on. Previously we were not computing the flattening coefficient in all the zones where we applied it. This also reuses the C++ flattening routine.

Note: this adds QPTOT to the primitive variable state so we can flattening on that.