Simplify enforce minimum density by always resetting to small_dens (#989)
When the density is < small_dens, we unconditionally reset it to small_dens. castro.density_reset_method is removed.

The methodology we chose for how to do a reset arguably mattered when we were relying on it for cleaning up bad advances that resulted in a negative density. But now we are relying on retries to handle this (#796). So enforce_min_density should only be cleaning up two cases: when a FillPatch generates a negative density (which is pretty unlikely), and when a reflux generates a negative density (possible, but we'll need to address that separately (#341).

This also avoids the issue discussed in #985 since now a reset is a fully zone-local operation.