Add fluxes consistently between simplified SDC and CTU (#1018)
At the end of the CTU hydro loop, we store the local Fab data from the fluxes and pressure into fluxes, mass_fluxes, and P_radial. This is done as an add because in general we subcycle the CTU advance, and we want the fluxes at the end of the timestep to be the sum of the fluxes from all subcycles. Then this data is reused later in the reflux.

Previously, for simplified SDC, we were copying the data instead of adding it, because we didn't want to add for each simplified SDC iteration, and when retry subcycling was added, it didn't work for simplified SDC. However we can now retry with simplified SDC, so we make it consistent with the CTU advance by adding fluxes as well, but only on the last SDC iteration.

In order to do this we modify Castro_advance_ctu.cpp slightly so that sdc_iters is directly modified on a retry (so that we can read its current value in the hydro advance).

We also correct the fact that P_radial was being stored inconsistently; we stored it always for non-Cartesian geometries, which was different from the logic we used for calculating the local pradial_fab update.