Fix lambda capture of dx in MGutils (#1039)
Device lambda capture of dx in the MGutils routines (apply_metric, etc.) requires that we use a GpuArray (geom.CellSizeArray()) rather than a raw pointer. The current code resulted in an illegal memory access error in CUDA when run with 2D Poisson gravity (and there were likely issues in diffusion in 2D as well). This issue was introduced in the port of this code from Fortran to C++ (#854).