Convert gravity test residual to C++ (#1041)
ca_test_residual is converted from Fortran to a C++ member of the Gravity class, test_residual. It is made dimension agnostic in the process.

Note: this function is only evaluated in DEBUG mode, and has no effect on the time evolution (it just prints a residual to stdout if gravity.v >= 2).