Remove remaining calls to amrex_filccn (#1048)
PR #837 removed the need for any problem setup to call amrex_filccn in its local bc_fill_nd.F90 routines, hypfill and denfill, because now we will always call the AMReX BC routines prior to entering hypfill or denfill. However that PR did not remove the calls to it from the problem setups that used it, noting that these calls are now redundant. These calls are now removed.