Remove restriction on CUDA that amr.blocking_factor be a multiple of 8 (#1049)
Though using a blocking factor >= 8 is recommended for performance reasons, it was previously required due to correctness reasons because of the use of an AMReX Fortran function, amrex_filccn. As noted in #1048, this function is no longer required due to recent changes in Castro (problems overriding bc_fill_nd.F90 or bc_ext_fill_nd.F90 do not need to provide an initial fill of the ghost zone data before implementing their specific boundary conditions; this is now done for them). Calling this function may now result in race conditions and correctness issues in the CUDA build, so it should be removed from any problem setups.