Don't call EOS on zones with rho = 0 (#1077)
ca_compute_avgpres is called for GR gravity. For monopole gravity we add contributions from multiple levels but mask out the contributions from coarse zones underlying fine levels. This means that in some calls to this function the state will be zeroed out. In that case we want to avoid calling the EOS as this generates NaNs (e.g. when evaluating 1 / rho).