Always use previous phi as guess for solve (#1078)
The Poisson solve has an option to not use the previous gravitational potential as a guess for the new solve (instead, a monopole solve is done and used to fill the domain). This is only used after initialization or restart. This option is now turned off, so that we always use the previous phi. At initialization we will pay a (one-time) larger cost to do the solve, but this should have no other effects.

I would like to remove make_radial_phi; for the most part we don't need it and it makes monopole gravity harder to maintain. If this bit is removed then the only remaining use of make_radial_phi is in the BCs for the Poisson solve, and there it can be replaced entirely with the multipole BCs.