Always refine on non-zero dxnuc in Detonation collision (#1114)
It seems to be the case for this problem that if we refine from the very beginning of the burning, even if we just subcycle the hydro on the fine levels with castro.reactions_max_solve_level=0, that helps avoid the spurious prompt ignition with a low base resolution. With the previous dxnuc_min, we were responding far too late to move the setup away from its course to ignition.