Copy additional data from old Castro level on regrid (#1122)
On a regrid where we're copying data from an existing level, we need to make sure we copy the class data relevant to post_timestep if a retry has occurred. This change copies some of these scalar parameters, and also fillpatches the old data in addition to the new data (if the old data is available in the state we're copying from).