use_post_step_regrid unconditionally triggers a post-step regrid (#898)
use_post_step_regrid previously would check if the set of tagged zones at the old time did not cover all zones tagged at the new time, and conditionally asked for a post-step regrid if so. However this was expensive on GPUs due to the serial nature of the algorithm (#887). Rather than sort this out, we just unconditionally regrid at the end of every step when the user uses this option. While regridding is expensive, the expectation is that this option is only used when the user expects their tagging criteria to be unconditionally satisfied at all times, and this comes with a performance tradeoff.