Selective clang format  (#1158)
This amends the C++ linter action so that it selects which problem(s) to compile depending on the files that have been changed in the PR. The action will fail on the first problem that fails to compile. An example of this working can be seen here: https://github.com/harpolea/Castro/runs/945245388.

Most of the work done to select the files and do the processing is done by https://github.com/AMReX-Astro/cpp-linter-action/blob/master/find_changed_files.py and https://github.com/AMReX-Astro/cpp-linter-action/blob/master/run_on_changed_files.py.