Allow problems to turn off use of GPU pragma (#1282)
If a problem is known to not require the GPU pragma script anymore, this can now be disabled by setting USE_GPU_PRAGMA = FALSE in its GNUmakefile. We implement this for Sedov and thus this closes #1203. Technically we have not eliminated all of the places in the Sedov build where pragma gpu is still in a compile path -- as mentioned there, the relevant places are set_problem_tags and then hypfill/denfill and ext_fill and ext_denfill. However, set_problem_tags does no work for Sedov, so there's no need for a GPU compile (it will just be a no-op on the CPU). A similar comment is true for the BC fill routines -- they only still do work at this point if we're using ambient BCs (hypfill/denfill) or HSE BCs (ext_fill/ext_denfill). So we should be able to eliminate the GPU build for those routines and trust that there is no performance impact because the loops that do any work never get evaluated.