only support USE_SIMPLIFIED_SDC with USE_CXX_REACTIONS=TRUE (#1294)
Now that we have wrappers that provide C++ interfaces for any pure-Fortran network, we can enforce that
simplified-SDC only works with a C++ integrator. This will make it easier for use to develop new features by not requiring us to port those features to both the C++ and Fortran code paths.