Apply Poisson gravity tagging restriction to all setups (#1302)
Poisson gravity doesn't have support for fine levels that touch the physical domain boundary (that is, the boundary is not either periodic or symmetric). This is enforced as an abort in the Poisson solver, but we can also do better by just preventing this from happening directly. This was already the case in wdmerger; this PR applies this across all problems that use Poisson gravity.