Add C++ interface for problems to override BCs (#1323)
A problem-facing hook, problem_bc_fill, is added in C++ for problems to set custom boundary conditions. The idea is that this should replace previous uses of overriding bc_fill_nd.F90. DustCollapse and Detonation are updated to the new scheme.

Note: we are not going to provide an equivalent of denfill, which would most likely be misused.