convert gamma_law_bubble to C++ (#1338)
The switches the initial model generation to be stored in the model_parser data structures.
It also removes the custom boundary conditions that were done previously to instead use
our general HSE BCs. This change will cause some differences in the output, but it should
be fine. The old BCs were essentially integrating the initial model into the BCs, so it was fixed
in time.