Include ghost zones when filling data on new level (#1344)
This addresses the issue that was unintentionally exposed by #1342. When we're creating a new fine level and filling data from the coarse level below it, we need to include the ghost zones in the fillpatch operation, or else they will be uninitialized and cause problems during the subsequent advance.

This can affect radiation and simplified-SDC problems.