Convert Rad2Tshock BCs to C++ (#1342)
Rad2Tshock's BCs are converted to C++. Since it was previously using EXT_DIR and then overriding with ext_fill, simply moving the BCs to problem_bc_fill.H is not sufficient, because that will still have the main Castro EXT_DIR routines be called, which is not what we want. So the inputs files are updated to use outflow BCs along the problem direction -- the outflow fill will get overwritten anyway.

This does cause a breaking change in the regression suite. Previously this problem was also using ext_denfill and filling the density on either end. We do not attempt to replace it, since ext_denfill is usually incorrect to use in the way it is used here. In this case, because this was being used for derives, it caused spurious tagging at either end of the domain, where the temperature saw a "gradient" due to a density-scaled value being located adjacent to the temperature-scaled value in the interior. In this new configuration, the spurious tagging at the domain edges is avoided.