convert double_bubble to C++ (#1343)
This changes the boundary conditions to use the centralized BCs and also puts the initial
model into the model_parser data structures.

The old BC integrated the initial model analytic profile into the ghost cells. For an isentropic atmosphere, this used the proper isentropic profile. Now we interpolate T into the ghost cells and integrate p via HSE. So it is not exactly the same as the analytic profile. I think it is good enough for a general HSE BC.