Convert get_omega to C++ (#1352)
get_omega is ported to C++ as an inline function. All current uses of get_omega in C++ (which were calling into Fortran) are now replaced with the new implementation.

Since no part of the rotation simulation capability remains in Fortran, the interface to Fortran is removed. The Fortran version of get_omega is left for now because two problem setups in Fortran still depend on it.

(This change does cause a slight issue for wdmerger, which modifies rotational_period in probinit. However the impact is minor and should have no outcome on the simulation result, and at any rate will be fixed when its initialization is converted to C++.)